added test and return {"data": null}
